added resusable components for auth pages
