+++ frontend/src/components/RegisterModal.jsx
@@ -15,59 +15,14 @@
  * @component
  */
 
-import { useState, useCallback, useMemo } from 'react';
+import { useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { register } from '../api/auth';
-import { useForm, useUniversities } from '../hooks';
+import { useRegisterForm } from '../hooks';
 import { BaseModal } from './ui';
-import FormInput from './FormInput';
-import FormButton from './FormButton';
-import NameInputPair from './NameInputPair';
+import RegisterFormContent from './RegisterFormContent';
 import TermsLink from './TermsLink';
 import { Alert, Divider } from './ui';
 import { BrainCircuitIcon, XIcon } from './icons';
-import {
-  extractEduSubdomain,
-  isValidRegistrationEmail,
-  isWhitelistedEmail,
-} from '../utils/email';
-
-/**
- * UniversityDetectionStatus Component
- *
- * Displays the university detection result based on the user's email.
- * Shows different states: detected, whitelisted, or not found.
- */
-function UniversityDetectionStatus({ detectedUniversity, isWhitelisted, onRequestUniversity }) {
-  if (detectedUniversity) {
-    return (
-      <Alert variant="success" title="University detected">
-        You will be enrolled in <strong>{detectedUniversity.name}</strong>
-      </Alert>
-    );
-  }
-
-  if (isWhitelisted) {
-    return (
-      <Alert variant="info" title="Whitelisted email detected">
-        You can create an account without a university affiliation
-      </Alert>
-    );
-  }
-
-  return (
-    <Alert variant="warning" title="University not found">
-      No university matches your email domain.{' '}
-      <button
-        type="button"
-        onClick={onRequestUniversity}
-        className="font-medium underline text-amber-600 hover:text-amber-700"
-      >
-        Request to add your university
-      </button>
-    </Alert>
-  );
-}
 
 /**
  * RegisterModal Component
@@ -79,116 +34,32 @@ function UniversityDetectionStatus({ detectedUniversity, isWhitelisted, onReques
  */
 export default function RegisterModal({ isOpen, onClose, onSwitchToLogin }) {
   const navigate = useNavigate();
-  const { data: universities = [], isLoading: loadingUniversities } = useUniversities();
-  const [detectedUniversity, setDetectedUniversity] = useState(null);
 
-  // Map of email domains to university objects for O(1) lookup
-  const universityDomainMap = useMemo(() => {
-    const map = new Map();
-    universities.forEach((uni) => {
-      if (uni.emailDomain) {
-        map.set(uni.emailDomain.toLowerCase(), uni);
-      }
-    });
-    return map;
-  }, [universities]);
-
-  // Find university by email (handles subdomains like cs.mit.edu -> mit)
-  const findUniversityByEmail = useCallback((email) => {
-    const subdomain = extractEduSubdomain(email);
-    if (!subdomain) return null;
-
-    let uni = universityDomainMap.get(subdomain);
-    if (uni) return uni;
-
-    // Try base domain for subdomains
-    if (subdomain.includes('.')) {
-      const baseDomain = subdomain.split('.').pop();
-      uni = universityDomainMap.get(baseDomain);
-      if (uni) return uni;
-    }
-
-    return null;
-  }, [universityDomainMap]);
-
-  const {
-    formData,
-    error,
-    setError,
-    loading,
-    handleChange,
-    handleSubmit,
-  } = useForm({
-    initialValues: { email: '', password: '', firstName: '', lastName: '' },
-
-    // Update university detection when email changes
-    onFieldChange: (name, value) => {
-      if (name === 'email' && universityDomainMap.size > 0) {
-        setDetectedUniversity(findUniversityByEmail(value));
-      }
-    },
-
-    // Validation before submission
-    validate: (data) => {
-      if (!isValidRegistrationEmail(data.email)) {
-        return 'Please use your university .edu email address';
-      }
-      const isWhitelisted = isWhitelistedEmail(data.email);
-      if (!detectedUniversity && !isWhitelisted) {
-        return 'No university found for your email domain. Please contact support if you believe this is an error.';
-      }
-      if (data.password.length < 6) {
-        return 'Password must be at least 6 characters';
-      }
-      return null;
-    },
-
-    // Submit registration
-    onSubmit: async (data) => {
-      const response = await register({
-        email: data.email.trim(),
-        password: data.password,
-        firstName: data.firstName.trim(),
-        lastName: data.lastName.trim(),
-      });
-
-      // Close modal and navigate to verification page
+  // Use shared register form hook
+  const registerForm = useRegisterForm({
+    onSuccess: ({ email, university }) => {
       onClose();
       navigate('/verify-email', {
         replace: true,
-        state: {
-          email: response.email || data.email,
-          university: response.university || detectedUniversity,
-        },
+        state: { email, university },
+      });
+    },
+    onRequestUniversity: ({ email, firstName, lastName }) => {
+      onClose();
+      navigate('/request-university', {
+        state: { email, firstName, lastName },
       });
     },
   });
 
-  // Navigate to university request flow
-  const handleRequestUniversity = () => {
-    if (!formData.firstName.trim()) {
-      setError('Please enter your first name');
-      return;
+  // Track previous open state to reset only when modal transitions from closed to open
+  const prevIsOpenRef = useRef(isOpen);
+  useEffect(() => {
+    if (isOpen && !prevIsOpenRef.current) {
+      registerForm.reset();
     }
-    if (!formData.lastName.trim()) {
-      setError('Please enter your last name');
-      return;
-    }
-
-    // Close modal and navigate to request university page
-    onClose();
-    navigate('/request-university', {
-      state: {
-        email: formData.email,
-        firstName: formData.firstName,
-        lastName: formData.lastName,
-      },
-    });
-  };
-
-  // Derived state
-  const isWhitelistedDomain = isWhitelistedEmail(formData.email);
-  const showUniversityStatus = formData.email && isValidRegistrationEmail(formData.email);
+    prevIsOpenRef.current = isOpen;
+  }, [isOpen, registerForm]);
 
   return (
     <BaseModal
@@ -207,6 +78,7 @@ export default function RegisterModal({ isOpen, onClose, onSwitchToLogin }) {
         >
           <XIcon className="h-5 w-5 text-muted-foreground hover:text-foreground" />
         </button>
+
         {/* Logo and header */}
         <div className="text-center mb-8">
           <div className="flex items-center justify-center mb-6">
@@ -224,59 +96,13 @@ export default function RegisterModal({ isOpen, onClose, onSwitchToLogin }) {
           </p>
         </div>
 
-        {error && (
+        {registerForm.error && (
           <Alert variant="error" className="mb-4">
-            {error}
+            {registerForm.error}
           </Alert>
         )}
 
-        <form className="space-y-4" onSubmit={handleSubmit}>
-          <NameInputPair
-            firstName={formData.firstName}
-            lastName={formData.lastName}
-            onChange={handleChange}
-            disabled={loading}
-          />
-
-          <FormInput
-            type="email"
-            name="email"
-            placeholder="Enter your .edu email *"
-            value={formData.email}
-            onChange={handleChange}
-            disabled={loading}
-            required
-          />
-
-          {showUniversityStatus && (
-            <UniversityDetectionStatus
-              detectedUniversity={detectedUniversity}
-              isWhitelisted={isWhitelistedDomain}
-              onRequestUniversity={handleRequestUniversity}
-            />
-          )}
-
-          <FormInput
-            type="password"
-            name="password"
-            placeholder="Create a password (min 6 characters) *"
-            value={formData.password}
-            onChange={handleChange}
-            disabled={loading}
-            required
-            minLength={6}
-          />
-
-          <FormButton
-            type="submit"
-            disabled={loading || loadingUniversities}
-            loading={loading}
-            loadingText="Creating account..."
-            className="w-full mt-6"
-          >
-            Create account
-          </FormButton>
-        </form>
+        <RegisterFormContent {...registerForm} />
 
         {/* Footer links */}
         <div className="mt-6 space-y-4">
@@ -286,6 +112,7 @@ export default function RegisterModal({ isOpen, onClose, onSwitchToLogin }) {
 
           <Divider>or</Divider>
 
+          {/* Switch to login MODAL */}
           <div className="text-center">
             <p className="text-sm text-muted-foreground">
               Already have an account?{' '}
@@ -307,4 +134,3 @@ export default function RegisterModal({ isOpen, onClose, onSwitchToLogin }) {
     </BaseModal>
   );
 }
-
